requirements.txt
aiogram==3.13.1
SQLAlchemy==2.0.35
asyncpg==0.29.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.2